split restore/backup commands because the commands are distinct.
